pyproject.toml
[project]
name = "microbiology-burmese-rag"
version = "0.1.0"
description = ""
authors = [
    {name = "zwe-htet-paing",email = "zwehtetpaing.works@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langsmith (>=0.2.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-core (>=0.3.74,<0.4.0)",
    "langchain-openai (>=0.3.30,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "lxml (>=6.0.0,<7.0.0)",
    "notebook (>=7.4.5,<8.0.0)",
    "streamlit (>=1.48.1,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
